agent/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "serdash-agent"
version = "0.1.0"
description = "Serdash metrics agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=42",
    "pyjwt[crypto]>=2.8",
    "requests>=2.31",
]

[project.scripts]
serdash-agent = "serdash_agent.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["serdash_agent*"]
